+++ NEWS.md
@@ -1,5 +1,6 @@
 # clusterProfiler 3.15.1
 
++ add `keytype` and `readable` info in `compareCluster` output (2019-12-02, Mon)
 + mv `compareClusterResult` class defintion to `DOSE` (2019-11-02, Sat)
 + mv `fortify`, `barplot` and  `dotplot` for `compareClusterResult` to `enrichplot`.